Introduce `cli-api` module
