+++ cli/build.gradle.kts
@@ -40,12 +40,15 @@ plugins {
 }
 
 dependencies {
+    implementation(Clikt.lib)
+    implementation(Flogger.lib)
+
+    implementation(project(":cli-api"))
     implementation(project(":api"))
     implementation(project(":compiler"))
     implementation(project(":codegen-java"))
     implementation(kotlin("reflect"))
-    implementation(Clikt.lib)
-    implementation(Flogger.lib)
+
     runtimeOnly(Flogger.Runtime.systemBackend)
 
     testImplementation(project(":test-env"))
