+++ cli/src/test/kotlin/io/spine/protodata/param/CliTest.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.cli
+package io.spine.protodata.param
 
 import com.github.ajalt.clikt.core.MissingOption
 import com.github.ajalt.clikt.core.UsageError
@@ -35,11 +35,12 @@ import io.spine.base.Time
 import io.spine.json.Json
 import io.spine.option.OptionsProto
 import io.spine.protobuf.AnyPacker
-import io.spine.protodata.cli.given.CustomOptionPlugin
-import io.spine.protodata.cli.given.CustomOptionRenderer
-import io.spine.protodata.cli.given.DefaultOptionsCounterPlugin
-import io.spine.protodata.cli.given.DefaultOptionsCounterRenderer
-import io.spine.protodata.cli.given.TestOptionProvider
+import io.spine.protodata.cli.Run
+import io.spine.protodata.param.given.CustomOptionPlugin
+import io.spine.protodata.param.given.CustomOptionRenderer
+import io.spine.protodata.param.given.DefaultOptionsCounterPlugin
+import io.spine.protodata.param.given.DefaultOptionsCounterRenderer
+import io.spine.protodata.param.given.TestOptionProvider
 import io.spine.protodata.cli.test.TestOptionsProto
 import io.spine.protodata.cli.test.TestProto
 import io.spine.protodata.test.ECHO_FILE
