+++ cli/src/test/kotlin/io/spine/protodata/param/given/CustomFieldView.kt
@@ -24,12 +24,36 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.cli
+package io.spine.protodata.param.given
 
-/**
- * Option names that are used in the help texts of other options.
- */
-public object Op {
-    public const val RENDERER: String = "--renderer"
-    public const val SOURCE_ROOT: String = "--source-root"
+import io.spine.core.External
+import io.spine.core.Subscribe
+import io.spine.core.Where
+import io.spine.protodata.FieldName
+import io.spine.protodata.event.FieldOptionDiscovered
+import io.spine.protodata.cli.test.CustomField
+import io.spine.protodata.plugin.View
+import io.spine.protodata.plugin.ViewRepository
+import io.spine.server.entity.alter
+import io.spine.server.route.EventRoute.withId
+import io.spine.server.route.EventRouting
+
+class CustomFieldView : View<FieldName, CustomField, CustomField.Builder>() {
+
+    @Subscribe
+    internal fun on(@External @Where(field = "option.name", equals = "custom")
+                        event: FieldOptionDiscovered
+    ) = alter {
+        field = event.field
+    }
+
+    class Repository : ViewRepository<FieldName, CustomFieldView, CustomField>() {
+
+        override fun setupEventRouting(routing: EventRouting<FieldName>) {
+            super.setupEventRouting(routing)
+            routing.route(FieldOptionDiscovered::class.java) { event, _ ->
+                withId(event.field)
+            }
+        }
+    }
 }
