+++ cli/src/test/kotlin/io/spine/protodata/param/given/DefaultOptionsCounterRenderer.kt
@@ -24,26 +24,29 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.cli.given
+package io.spine.protodata.param.given
 
-import io.spine.protodata.cli.test.CustomField
-import io.spine.tools.code.CommonLanguages
+import io.spine.protodata.cli.test.DefaultOptionsCounter
 import io.spine.protodata.renderer.Renderer
 import io.spine.protodata.renderer.SourceFileSet
 import io.spine.server.query.select
+import io.spine.tools.code.CommonLanguages
 import kotlin.io.path.Path
 
-class CustomOptionRenderer : Renderer(CommonLanguages.any) {
+class DefaultOptionsCounterRenderer : Renderer(CommonLanguages.any) {
 
     companion object {
-        const val FILE_NAME = "custom_fields.csv"
+        const val FILE_NAME = "default_opts_counted.txt"
     }
 
     override fun render(sources: SourceFileSet) {
-        val customFields = select<CustomField>().all()
+        val counters = select<DefaultOptionsCounter>().all()
         sources.createFile(
             Path(FILE_NAME),
-            customFields.joinToString(separator = ",") { it.field.value }
+            counters.joinToString(separator = ",") {
+                it.timestampInFutureEncountered.toString() + ", " +
+                        it.requiredFieldForTestEncountered.toString()
+            }
         )
     }
 }
